+++ R/ReactiveObject.S3.r
@@ -7,7 +7,7 @@
 #' @details
 #' Instances of this class are implicitly created when calling 
 #' \code{\link[reactr]{setReactiveS3}} and stored in the registry. Objects of 
-#' this class can be thought of as the "hidden parts" of the reactive objects 
+#' this class can be thought of as the "invisible parts" of the reactive objects 
 #' of whom actually only field \code{.value} is visible to the user or other 
 #' functions consuming the reactive object.
 #' 
@@ -161,7 +161,6 @@ ReactiveObject.S3 <- function(
   
   ## Conditions and status messages //
   condition = NULL
-  
 ) {
   if (!missing(.x)) {
     class(.x) <- c("ReactiveObject.S3", class(.x))