+++ R/copyReactive.r
@@ -6,10 +6,10 @@
 #' 
 #' @details
 #' Takes the essential information of a reactive object (fields \code{value}, 
-#' \code{refs_pull}, \code{refs_push} and \code{funct} of the hidden instance 
+#' \code{refs_pull}, \code{refs_push} and \code{funct} of the invisible instance 
 #' of class \code{\link[reactr]{ReactiveObject.S3}} and creates a new reactive 
 #' object of name \code{id} in \code{where} with that information. 
-#' The important thing to note is that the two objects (and their hidden objects)
+#' The important thing to note is that the two objects (and their invisible objects)
 #' are not connected via any environments (pass-by-reference) as would be the 
 #' case if the copy was carried out by a mere call to \code{\link[base]{<-}}.
 #'