+++ inst/examples/unsetReactive.r
@@ -30,8 +30,8 @@ identical(x_2, x_1)
 ## --> reactive relationship re-established again
 
 ## Clean up //
-removeReactive("x_1")
-removeReactive("x_2")
+rmReactive("x_1")
+rmReactive("x_2")
 
 ##------------------------------------------------------------------------------  
 ## Strict get behavior level 1 //  
@@ -50,8 +50,8 @@ x_2
 ## --> reactive relationship restored
 
 ## Clean up //
-removeReactive("x_1")
-removeReactive("x_2")
+rmReactive("x_1")
+rmReactive("x_2")
 
 ##------------------------------------------------------------------------------  
 ## Strict get behavior level 2 //  
@@ -70,7 +70,7 @@ x_2
 ## --> reactive relationship restored
 
 ## Clean up //
-removeReactive("x_1")
-removeReactive("x_2")
+rmReactive("x_1")
+rmReactive("x_2")
 
 }